Figma Project - Assignment 5 Sem2 UX/UI design
Timeline of Game Console Evolution – Figma Project

In this assignment, you will create a wireframe for a fun and informative website that effectively explains a chosen topic to users while incorporating key UX/UI design principles. This wireframe is the foundation for Part 2 of the assignment, where you will develop an interactive prototype using Figma. 

Console Evolution - Discover the evolution of video game From the 1950s to today!